requirements.txt
streamlit==1.30.0
PyPDF2==3.0.1
pillow==10.1.0
pytesseract==0.3.10
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
nltk==3.8.1
tokenizers>=0.13.2 